Added History Deleted Counter

+++ background.js
@@ -1,3 +1,31 @@
+const SECOND = 1000;
+const MINUTE = 60 * SECOND;
+const HOUR = 60 * MINUTE;
+const DAY = 24 * HOUR;
+const DELAYINMINUTES = 60;
+
+//Keep History for X amount of days
+function deleteOldHistory() {
+	browser.history.deleteRange({startTime: 0, endTime: Date.now() - (DAY*daysToKeep)} , function() {
+			//console.log("History cleared");
+	});
+
+}
+
+//Creates an alarm that clears history every hour
+function createOldHistoryAlarm() {
+	browser.alarms.create("historyAutoDeleteAlarm", {
+		periodInMinutes: DELAYINMINUTES
+	});
+}
+
+//Deletes the alarm
+function deleteOldHistoryAlarm() {
+    //console.log("Deleted");
+   	browser.alarms.clear("historyAutoDeleteAlarm");	
+}
+
+
 //Logs the error
 function onError(error) {
 	console.error(`Error: ${error}`);
@@ -5,11 +33,18 @@ function onError(error) {
 
 
 //Returns the host name of the url. Etc. "https://en.wikipedia.org/wiki/Cat" becomes en.wikipedia.org
-function get_hostname(url) {
+function getHostname(url) {
     var hostname = new URL(url).hostname;
     // Strip "www." if the URL starts with it.
     hostname = hostname.replace(/^www\./, '');
     return hostname;
+}
+
+function isAWebpage(URL) {
+	if(URL.match(/^http:/) || URL.match(/^https:/)) {
+		return true;
+	}
+	return false;
 }
 
 //See if the set has the url
@@ -50,38 +85,96 @@ function clearURL() {
 function onVisited(historyItem) {
 	if (historyItem.url) {
 		var currentUrl = historyItem.url;
-		var currentHostUrl = get_hostname(currentUrl);
+		var currentHostUrl = getHostname(currentUrl);
 		//console.log(currentUrl);
 		//console.log("Host: " + currentHostUrl);
 		if(hasHost(currentHostUrl)) {
 			var deletingUrl = browser.history.deleteUrl({url: currentUrl});
 			deletingUrl.then(function() {
-				console.log(currentURL + " deleted from history");
+				//console.log(currentURL + " deleted from history");
 			});
 		}
 	}
 
 }
 
-//The set of urls
-var urlsToRemove;
-
-
-var getURLSFromStorage = browser.storage.local.get("URLS");
-getURLSFromStorage.then(function(item) {
-	urlsToRemove = new Set(item.URLS);
-}).catch(onError);
-
-browser.history.onVisited.addListener(onVisited);
-
-//Logic that controls when to disable the browser action
-function isAWebpage(URL) {
-	if(URL.match(/^http:/) || URL.match(/^https:/)) {
-		return true;
+//Increment the counter and store the counter to local after 1 minute
+function incrementCounter() {
+	if (statLoggingSetting == true) {
+		historyDeletedCounterTotal++;
+		historyDeletedCounter++;
+		browser.alarms.create("storeCounterToLocalAlarm", {
+			delayInMinutes: 1
+		});		
+		//console.log(historyDeletedCounterTotal);
 	}
-	return false;
+
 }
 
+//Resets the counter
+function resetCounter() {
+	browser.storage.local.set({historyDeletedCounterTotal: 0});
+	historyDeletedCounterTotal = 0;
+	historyDeletedCounter = 0;
+}
+
+//Stores the total history entries deleted to local
+function storeCounterToLocal() {
+	browser.storage.local.set({historyDeletedCounterTotal: historyDeletedCounterTotal});
+}
+
+//Set the variables from local storage
+function initializeVariables() {
+	var getLocal = browser.storage.local.get();
+	getLocal.then(function(items) {
+		if(items.daysToKeep == null) {
+			setDefaults();
+		}
+		urlsToRemove = new Set(items.URLS);
+		daysToKeep = items.daysToKeep;
+		historyDeletedCounterTotal = items.historyDeletedCounterTotal;
+		keepHistorySetting = items.keepHistorySetting;
+		statLoggingSetting = items.statLoggingSetting;
+		//console.log(historyDeletedCounterTotal + ", " + keepHistorySetting + ", " + statLoggingSetting);
+	}).catch(onError);
+}
+
+//Create some objects based on the values from initializing the variables
+function initializeObjects() {
+	if(keepHistorySetting != null && keepHistorySetting == true) {
+		deleteOldHistory();
+		createOldHistoryAlarm();
+	}
+
+}
+
+//Set the defaults if nothing appears in local storage
+function setDefaults() {
+	console.log("Set defaults");
+	browser.storage.local.set({
+		daysToKeep: 60,
+		historyDeletedCounterTotal: 0,
+		keepHistorySetting: false,
+		statLoggingSetting: true
+	});
+}
+//The set of urls
+var urlsToRemove;
+//Numeric values
+var daysToKeep;
+var historyDeletedCounterTotal;
+var historyDeletedCounter = 0;
+//Boolean values
+var keepHistorySetting;
+var statLoggingSetting;
+initializeVariables();
+initializeObjects();
+
+//History Event handler
+browser.history.onVisited.addListener(onVisited);
+browser.history.onVisitRemoved.addListener(incrementCounter);
+
+//Logic that controls when to disable the browser action
 browser.tabs.onUpdated.addListener(function(tabId, changeInfo, tab) {
 	browser.windows.getCurrent(function(windowInfo) {
 		if (!isAWebpage(tab.url) || windowInfo.incognito) {
@@ -95,47 +188,15 @@ browser.tabs.onUpdated.addListener(function(tabId, changeInfo, tab) {
 
 });
 
-/*
-Keep History for X days Logic
-*/
-
-const SECOND = 1000;
-const MINUTE = 60 * SECOND;
-const HOUR = 60 * MINUTE;
-const DAY = 24 * HOUR;
-const DELAYINMINUTES = 60;
-
-//Keep History for X amount of days
-function deleteOldHistory() {
-	browser.storage.local.get("daysToKeep", function(results) {
-		browser.history.deleteRange({startTime: 0, endTime: Date.now() - DAY*results.daysToKeep} , function() {
-			//console.log("History cleared");
-		});
-	});
-
-}
-
-//Creates an alarm that clears history every hour
-function createOldHistoryAlarm() {
-	browser.alarms.create("historyAutoDeleteAlarm", {
-		periodInMinutes: DELAYINMINUTES
-	});
-}
-
-//Deletes the alarm
-function deleteOldHistoryAlarm() {
-    //console.log("Deleted");
-   	browser.alarms.clear("historyAutoDeleteAlarm");	
-}
-
-browser.alarms.onAlarm.addListener(alarmInfo => {
-	deleteOldHistory();
-});
-
-//On startup clear history and create the alarm if the user selects this option
-browser.storage.local.get("keepHistorySetting", function(results) {
-	if(results.keepHistorySetting != null && results.keepHistorySetting == true) {
-		deleteOldHistory();
-		createOldHistoryAlarm();
+//Alarm event handler
+browser.alarms.onAlarm.addListener(function (alarmInfo) {
+	console.log(alarmInfo.name);
+	if(alarmInfo.name == "historyAutoDeleteAlarm") {
+		deleteOldHistory();		
 	}
+	if(alarmInfo.name == "storeCounterToLocalAlarm") {
+		storeCounterToLocal();
+	}
+
 });
+
